move metrics-definition-types to common
